Remove colors Gold, Silver, Dark Blue add colors for rebranded buttons - white, blue, black add border for white color button
