+++ Demo/src/main/java/com/paypal/android/ui/paypal/PayPalButtonColorOptionList.kt
@@ -1,22 +1,22 @@
-package com.paypal.android.ui.paypalbuttons
+package com.paypal.android.ui.paypal
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.paypal.android.R
-import com.paypal.android.paymentbuttons.PayPalCreditButtonColor
+import com.paypal.android.paymentbuttons.PayPalButtonColor
 import com.paypal.android.uishared.components.OptionList
 
 @Composable
-fun PayPalCreditButtonColorOptionList(
-    selectedOption: PayPalCreditButtonColor,
-    onSelection: (PayPalCreditButtonColor) -> Unit
+fun PayPalButtonColorOptionList(
+    selectedOption: PayPalButtonColor,
+    onSelection: (PayPalButtonColor) -> Unit
 ) {
     OptionList(
         title = stringResource(id = R.string.pay_pal_button_color),
-        options = PayPalCreditButtonColor.values().map { it.name },
+        options = PayPalButtonColor.entries.map { it.name },
         selectedOption = selectedOption.name,
         onSelectedOptionChange = { option ->
-            onSelection(PayPalCreditButtonColor.valueOf(option))
+            onSelection(PayPalButtonColor.valueOf(option))
         }
     )
 }
