+++ Demo/src/main/java/com/paypal/android/ui/paypal/PayPalButtonLabelOptionList.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paypalbuttons
+package com.paypal.android.ui.paypal
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
@@ -13,7 +13,7 @@ fun PayPalButtonLabelOptionList(
 ) {
     OptionList(
         title = stringResource(id = R.string.pay_pal_button_label),
-        options = PayPalButtonLabel.values().map { it.name },
+        options = PayPalButtonLabel.entries.map { it.name },
         selectedOption = selectedOption.name,
         onSelectedOptionChange = { option ->
             onSelection(PayPalButtonLabel.valueOf(option))
