+++ Demo/src/main/java/com/paypal/android/ui/paypal/PayPalCreditButtonColorOptionList.kt
@@ -1,22 +1,22 @@
-package com.paypal.android.ui.paypalbuttons
+package com.paypal.android.ui.paypal
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.paypal.android.R
-import com.paypal.android.paymentbuttons.PayPalButtonColor
+import com.paypal.android.paymentbuttons.PayPalCreditButtonColor
 import com.paypal.android.uishared.components.OptionList
 
 @Composable
-fun PayPalButtonColorOptionList(
-    selectedOption: PayPalButtonColor,
-    onSelection: (PayPalButtonColor) -> Unit
+fun PayPalCreditButtonColorOptionList(
+    selectedOption: PayPalCreditButtonColor,
+    onSelection: (PayPalCreditButtonColor) -> Unit
 ) {
     OptionList(
         title = stringResource(id = R.string.pay_pal_button_color),
-        options = PayPalButtonColor.values().map { it.name },
+        options = PayPalCreditButtonColor.entries.map { it.name },
         selectedOption = selectedOption.name,
         onSelectedOptionChange = { option ->
-            onSelection(PayPalButtonColor.valueOf(option))
+            onSelection(PayPalCreditButtonColor.valueOf(option))
         }
     )
 }
