+++ Demo/src/main/java/com/paypal/android/ui/paypal/PaymentButtonShapeOptionList.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paypalbuttons
+package com.paypal.android.ui.paypal
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
@@ -13,7 +13,7 @@ fun PaymentButtonShapeOptionList(
 ) {
     OptionList(
         title = stringResource(id = R.string.pay_pal_button_shape),
-        options = PaymentButtonShape.values().map { it.name },
+        options = PaymentButtonShape.entries.map { it.name },
         selectedOption = selectedOption.name,
         onSelectedOptionChange = { option ->
             onSelection(PaymentButtonShape.valueOf(option))
