+++ Demo/src/main/res/layout/pay_later_button_test_layout.xml
@@ -37,18 +37,6 @@
             android:layout_margin="8dp"
             app:paypal_color="blue" />
 
-        <com.paypal.android.paymentbuttons.PayPalButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            app:paypal_color="gold" />
-
-        <com.paypal.android.paymentbuttons.PayPalButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            app:paypal_color="silver" />
-
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -65,19 +53,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="8dp"
-            app:paylater_color="gold" />
-
-        <com.paypal.android.paymentbuttons.PayLaterButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
             app:paylater_color="blue" />
-
-        <com.paypal.android.paymentbuttons.PayLaterButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            app:paylater_color="silver" />
 
         <com.paypal.android.paymentbuttons.PayLaterButton
             android:layout_width="match_parent"
@@ -96,12 +72,6 @@
             android:layout_height="wrap_content"
             android:layout_margin="8dp"
             app:paypal_credit_color="black" />
-
-        <com.paypal.android.paymentbuttons.PayPalCreditButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            app:paypal_credit_color="dark_blue" />
 
     </LinearLayout>
 </ScrollView>
