+++ PaymentButtons/src/main/res/color/paypal_black.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#2C2E2F" android:state_pressed="true" />
-    <item android:color="#000000" />
+    <item android:color="@color/neutral_black_600" android:state_pressed="true" />
+    <item android:color="@color/neutral_black" />
 </selector>
