+++ PaymentButtons/src/main/res/color/paypal_blue.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#0063AD" android:state_pressed="true" />
-    <item android:color="#0070BA" />
+    <item android:color="@color/blue_400" android:state_pressed="true" />
+    <item android:color="@color/blue_300" />
 </selector>
