--- PaymentButtons/src/main/res/color/paypal_dark_blue.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#14469D" android:state_pressed="true" />
-    <item android:color="#073990" />
-</selector>
