--- PaymentButtons/src/main/res/color/paypal_silver.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#E2E2E2" android:state_pressed="true" />
-    <item android:color="#EEEEEE" />
-</selector>
