+++ PaymentButtons/src/main/res/color/paypal_white.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="#EAECED" android:state_pressed="true" />
-    <item android:color="#FFFFFF" />
+    <item android:color="@color/neutral_white_pressed" android:state_pressed="true" />
+    <item android:color="@color/neutral_white" />
 </selector>
