+++ PaymentButtons/src/main/res/values/dimens.xml
@@ -9,6 +9,7 @@
     <dimen name="paypal_payment_button_text_margin">@dimen/margin_8x</dimen>
     <!-- The stroke used for white buttons to help them standout on light backgrounds. -->
     <dimen name="paypal_payment_button_stroke_width">1dp</dimen>
+    <dimen name="button_white_border_width">4dp</dimen>
     <!-- 48dp is the minimum height of buttons to account for accessibility requirements -->
     <dimen name="paypal_payment_button_min_height">48dp</dimen>
     <dimen name="paypal_payment_button_min_height_large">55dp</dimen>
